requirements.txt
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dominate==2.3.4
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.5
requests==2.20.0
SQLAlchemy==1.2.12
urllib3==1.24
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
